+++ django-data/image/submissions/forms.py
@@ -22,6 +22,9 @@ from excel.helpers import ExcelTemplateReader
 
 class SubmissionFormMixin():
     def clean(self):
+        # test if I have a submission with the provided data
+        self.check_submission_exists()
+
         # I can call this method without providing a 'uploaded file'
         # (for instance, when omitting uploaded file)
         if "uploaded_file" in self.cleaned_data:
@@ -39,6 +42,29 @@ class SubmissionFormMixin():
             if ("datasource_type" in self.cleaned_data and
                     self.cleaned_data["datasource_type"] == TEMPLATE_TYPE):
                 self.check_template_file()
+
+    def check_submission_exists(self):
+        """Test if I already have a submission with the same data"""
+
+        # get unique attributes
+        unique_together = Submission._meta.unique_together[0]
+
+        # get submitted attributes
+        data = {key: self.cleaned_data.get(key) for key in unique_together}
+
+        # ovverride owner attribute
+        data['owner'] = self.request.user
+
+        # test for a submission object with the same attributes
+        if Submission.objects.filter(**data).exists():
+            msg = (
+                "Error: There is already a submission with the same "
+                "attributes. Please change one of the following: "
+                "Gene bank name, Gene bank country, Data source type and "
+                "Data source version")
+
+            # raising an exception:
+            raise forms.ValidationError(msg, code='invalid')
 
     def check_file_encoding(self):
         uploaded_file = self.cleaned_data['uploaded_file']